+++ packages/nodes-base/nodes/Postgres/v2/actions/router.ts
@@ -3,9 +3,9 @@ import { NodeExecutionOutput, NodeOperationError } from 'n8n-workflow';
 
 import * as database from './database/Database.resource';
 import type { PostgresType } from './node.type';
+import { configurePostgres } from '../../transport';
 import type { PostgresNodeCredentials, PostgresNodeOptions } from '../helpers/interfaces';
 import { configureQueryRunner } from '../helpers/utils';
-import { configurePostgres } from '../transport';
 
 export async function router(this: IExecuteFunctions): Promise<INodeExecutionData[][]> {
 	let returnData: INodeExecutionData[] = [];
