+++ packages/nodes-base/nodes/Postgres/v2/methods/listSearch.ts
@@ -1,7 +1,7 @@
 import type { ILoadOptionsFunctions, INodeListSearchResult } from 'n8n-workflow';
 
 import type { PostgresNodeCredentials } from '../helpers/interfaces';
-import { configurePostgres } from '../transport';
+import { configurePostgres } from '../../transport';
 
 export async function schemaSearch(this: ILoadOptionsFunctions): Promise<INodeListSearchResult> {
 	const credentials = await this.getCredentials<PostgresNodeCredentials>('postgres');
