+++ packages/nodes-base/nodes/Postgres/v2/methods/resourceMapping.ts
@@ -2,7 +2,7 @@ import type { ILoadOptionsFunctions, ResourceMapperFields, FieldType } from 'n8n
 
 import type { PostgresNodeCredentials } from '../helpers/interfaces';
 import { getEnumValues, getEnums, getTableSchema, uniqueColumns } from '../helpers/utils';
-import { configurePostgres } from '../transport';
+import { configurePostgres } from '../../transport';
 
 const fieldTypeMapping: Partial<Record<FieldType, string[]>> = {
 	string: ['text', 'varchar', 'character varying', 'character', 'char'],
